functions/mailChimp/requirements.txt
google-cloud-secret-manager~=2.2
mailchimp3~=3.0